refactor(enterprise/coderd): move group AI budget handlers into aibridge.go
